Fix warnings in testdependencies

+++ tests/integration/testdependencies/dependency.c
@@ -1,10 +1,12 @@
 #include "dependency.h"
 #include <malloc.h>
+#include <stdlib.h>
+#include <stdint.h>
 
 int dep_run()
 {
 #if defined(__GLIBC_PREREQ) && __GLIBC_PREREQ(2, 17)
-    return (int)secure_getenv("NON_EXISTING_ENV_VARIABLE");
+    return (int)(intptr_t)secure_getenv("NON_EXISTING_ENV_VARIABLE");
 #elif defined(__GLIBC_PREREQ) && __GLIBC_PREREQ(2, 10)
     return malloc_info(0, stdout);
 #else
